Improve automation of regexp generation
Instead of having one script that generates .js files in scripts/unicode/, and then another script that consults those to build lib/regexes.js, we can just build regexes.js directly. This ensures that regexes.js will be updated as part of the prepublish process.
