+++ scripts/.eslintrc.json
@@ -1,0 +1,7 @@
+{
+  "rules": {
+    "no-process-env": "off",
+    "no-process-exit": "off",
+    "no-console": "off"
+  }
+}
