--- scripts/unicode/Canonical_Combining_Class.js
@@ -1,6 +1,0 @@
-// Generated from https://unicode.org/Public/15.0.0/ucd/extracted/DerivedCombiningClass.txt
-// $ node scripts/generateUnicodeCodePoints.js DerivedCombiningClass.txt 9
-
-module.exports = {
-  Virama: [2381,2509,2637,2765,2893,3021,3149,3277,3387,3388,3405,3530,3642,3770,3972,4153,4154,5908,5909,5940,6098,6752,6980,7082,7083,7154,7155,11647,43014,43052,43204,43347,43456,43766,44013,68159,69702,69744,69759,69817,69939,69940,70080,70197,70378,70477,70722,70850,71103,71231,71350,71467,71737,71997,71998,72160,72244,72263,72345,72767,73028,73029,73111,73537,73538]
-};
