--- scripts/unicode/Joining_Type.js
@@ -1,9 +1,0 @@
-// Generated from https://unicode.org/Public/15.0.0/ucd/extracted/DerivedJoiningType.txt
-// $ node scripts/generateUnicodeCodePoints.js DerivedJoiningType.txt L R D T
-
-module.exports = {
-  D: [1568,1574,1576,1578,1579,1580,1581,1582,1587,1588,1589,1590,1591,1592,1593,1594,1595,1596,1597,1598,1599,1601,1602,1603,1604,1605,1606,1607,1609,1610,1646,1647,1656,1657,1658,1659,1660,1661,1662,1663,1664,1665,1666,1667,1668,1669,1670,1671,1690,1691,1692,1693,1694,1695,1696,1697,1698,1699,1700,1701,1702,1703,1704,1705,1706,1707,1708,1709,1710,1711,1712,1713,1714,1715,1716,1717,1718,1719,1720,1721,1722,1723,1724,1725,1726,1727,1729,1730,1740,1742,1744,1745,1786,1787,1788,1791,1810,1811,1812,1818,1819,1820,1821,1823,1824,1825,1826,1827,1828,1829,1830,1831,1833,1835,1837,1838,1870,1871,1872,1873,1874,1875,1876,1877,1878,1879,1880,1884,1885,1886,1887,1888,1889,1890,1891,1892,1893,1894,1895,1896,1897,1898,1901,1902,1903,1904,1906,1909,1910,1911,1914,1915,1916,1917,1918,1919,1994,1995,1996,1997,1998,1999,2000,2001,2002,2003,2004,2005,2006,2007,2008,2009,2010,2011,2012,2013,2014,2015,2016,2017,2018,2019,2020,2021,2022,2023,2024,2025,2026,2113,2114,2115,2116,2117,2120,2122,2123,2124,2125,2126,2127,2128,2129,2130,2131,2133,2144,2146,2147,2148,2149,2152,2182,2185,2186,2187,2188,2189,2208,2209,2210,2211,2212,2213,2214,2215,2216,2217,2223,2224,2227,2228,2229,2230,2231,2232,2234,2235,2236,2237,2238,2239,2240,2241,2242,2243,2244,2245,2246,2247,2248,6151,6176,6177,6178,6179,6180,6181,6182,6183,6184,6185,6186,6187,6188,6189,6190,6191,6192,6193,6194,6195,6196,6197,6198,6199,6200,6201,6202,6203,6204,6205,6206,6207,6208,6209,6210,6211,6212,6213,6214,6215,6216,6217,6218,6219,6220,6221,6222,6223,6224,6225,6226,6227,6228,6229,6230,6231,6232,6233,6234,6235,6236,6237,6238,6239,6240,6241,6242,6243,6244,6245,6246,6247,6248,6249,6250,6251,6252,6253,6254,6255,6256,6257,6258,6259,6260,6261,6262,6263,6264,6279,6280,6281,6282,6283,6284,6285,6286,6287,6288,6289,6290,6291,6292,6293,6294,6295,6296,6297,6298,6299,6300,6301,6302,6303,6304,6305,6306,6307,6308,6309,6310,6311,6312,6314,43072,43073,43074,43075,43076,43077,43078,43079,43080,43081,43082,43083,43084,43085,43086,43087,43088,43089,43090,43091,43092,43093,43094,43095,43096,43097,43098,43099,43100,43101,43102,43103,43104,43105,43106,43107,43108,43109,43110,43111,43112,43113,43114,43115,43116,43117,43118,43119,43120,43121,68288,68289,68290,68291,68292,68307,68308,68309,68310,68312,68313,68314,68315,68316,68318,68319,68320,68331,68332,68333,68334,68480,68482,68486,68487,68488,68490,68491,68493,68496,68525,68526,68865,68866,68867,68868,68869,68870,68871,68872,68873,68874,68875,68876,68877,68878,68879,68880,68881,68882,68883,68884,68885,68886,68887,68888,68889,68890,68891,68892,68893,68894,68895,68896,68897,68899,69424,69425,69426,69428,69429,69430,69431,69432,69433,69434,69435,69436,69437,69438,69439,69440,69441,69442,69443,69444,69457,69458,69459,69488,69489,69490,69491,69494,69495,69496,69497,69498,69499,69500,69501,69502,69503,69504,69505,69552,69554,69555,69560,69563,69564,69566,69567,69569,69572,69578,125184,125185,125186,125187,125188,125189,125190,125191,125192,125193,125194,125195,125196,125197,125198,125199,125200,125201,125202,125203,125204,125205,125206,125207,125208,125209,125210,125211,125212,125213,125214,125215,125216,125217,125218,125219,125220,125221,125222,125223,125224,125225,125226,125227,125228,125229,125230,125231,125232,125233,125234,125235,125236,125237,125238,125239,125240,125241,125242,125243,125244,125245,125246,125247,125248,125249,125250,125251],
-  R: [1570,1571,1572,1573,1575,1577,1583,1584,1585,1586,1608,1649,1650,1651,1653,1654,1655,1672,1673,1674,1675,1676,1677,1678,1679,1680,1681,1682,1683,1684,1685,1686,1687,1688,1689,1728,1731,1732,1733,1734,1735,1736,1737,1738,1739,1741,1743,1746,1747,1749,1774,1775,1808,1813,1814,1815,1816,1817,1822,1832,1834,1836,1839,1869,1881,1882,1883,1899,1900,1905,1907,1908,1912,1913,2112,2118,2119,2121,2132,2134,2135,2136,2151,2153,2154,2160,2161,2162,2163,2164,2165,2166,2167,2168,2169,2170,2171,2172,2173,2174,2175,2176,2177,2178,2190,2218,2219,2220,2222,2225,2226,2233,68293,68295,68297,68298,68302,68303,68304,68305,68306,68317,68321,68324,68335,68481,68483,68484,68485,68489,68492,68494,68495,68497,68521,68522,68523,68524,68898,69427,69460,69492,69493,69556,69557,69558,69561,69562,69565,69570,69571,69577],
-  L: [43122,68301,68311,68864,69579],
-  T: [173,768,769,770,771,772,773,774,775,776,777,778,779,780,781,782,783,784,785,786,787,788,789,790,791,792,793,794,795,796,797,798,799,800,801,802,803,804,805,806,807,808,809,810,811,812,813,814,815,816,817,818,819,820,821,822,823,824,825,826,827,828,829,830,831,832,833,834,835,836,837,838,839,840,841,842,843,844,845,846,847,848,849,850,851,852,853,854,855,856,857,858,859,860,861,862,863,864,865,866,867,868,869,870,871,872,873,874,875,876,877,878,879,1155,1156,1157,1158,1159,1160,1161,1425,1426,1427,1428,1429,1430,1431,1432,1433,1434,1435,1436,1437,1438,1439,1440,1441,1442,1443,1444,1445,1446,1447,1448,1449,1450,1451,1452,1453,1454,1455,1456,1457,1458,1459,1460,1461,1462,1463,1464,1465,1466,1467,1468,1469,1471,1473,1474,1476,1477,1479,1552,1553,1554,1555,1556,1557,1558,1559,1560,1561,1562,1564,1611,1612,1613,1614,1615,1616,1617,1618,1619,1620,1621,1622,1623,1624,1625,1626,1627,1628,1629,1630,1631,1648,1750,1751,1752,1753,1754,1755,1756,1759,1760,1761,1762,1763,1764,1767,1768,1770,1771,1772,1773,1807,1809,1840,1841,1842,1843,1844,1845,1846,1847,1848,1849,1850,1851,1852,1853,1854,1855,1856,1857,1858,1859,1860,1861,1862,1863,1864,1865,1866,1958,1959,1960,1961,1962,1963,1964,1965,1966,1967,1968,2027,2028,2029,2030,2031,2032,2033,2034,2035,2045,2070,2071,2072,2073,2075,2076,2077,2078,2079,2080,2081,2082,2083,2085,2086,2087,2089,2090,2091,2092,2093,2137,2138,2139,2200,2201,2202,2203,2204,2205,2206,2207,2250,2251,2252,2253,2254,2255,2256,2257,2258,2259,2260,2261,2262,2263,2264,2265,2266,2267,2268,2269,2270,2271,2272,2273,2275,2276,2277,2278,2279,2280,2281,2282,2283,2284,2285,2286,2287,2288,2289,2290,2291,2292,2293,2294,2295,2296,2297,2298,2299,2300,2301,2302,2303,2304,2305,2306,2362,2364,2369,2370,2371,2372,2373,2374,2375,2376,2381,2385,2386,2387,2388,2389,2390,2391,2402,2403,2433,2492,2497,2498,2499,2500,2509,2530,2531,2558,2561,2562,2620,2625,2626,2631,2632,2635,2636,2637,2641,2672,2673,2677,2689,2690,2748,2753,2754,2755,2756,2757,2759,2760,2765,2786,2787,2810,2811,2812,2813,2814,2815,2817,2876,2879,2881,2882,2883,2884,2893,2901,2902,2914,2915,2946,3008,3021,3072,3076,3132,3134,3135,3136,3142,3143,3144,3146,3147,3148,3149,3157,3158,3170,3171,3201,3260,3263,3270,3276,3277,3298,3299,3328,3329,3387,3388,3393,3394,3395,3396,3405,3426,3427,3457,3530,3538,3539,3540,3542,3633,3636,3637,3638,3639,3640,3641,3642,3655,3656,3657,3658,3659,3660,3661,3662,3761,3764,3765,3766,3767,3768,3769,3770,3771,3772,3784,3785,3786,3787,3788,3789,3790,3864,3865,3893,3895,3897,3953,3954,3955,3956,3957,3958,3959,3960,3961,3962,3963,3964,3965,3966,3968,3969,3970,3971,3972,3974,3975,3981,3982,3983,3984,3985,3986,3987,3988,3989,3990,3991,3993,3994,3995,3996,3997,3998,3999,4000,4001,4002,4003,4004,4005,4006,4007,4008,4009,4010,4011,4012,4013,4014,4015,4016,4017,4018,4019,4020,4021,4022,4023,4024,4025,4026,4027,4028,4038,4141,4142,4143,4144,4146,4147,4148,4149,4150,4151,4153,4154,4157,4158,4184,4185,4190,4191,4192,4209,4210,4211,4212,4226,4229,4230,4237,4253,4957,4958,4959,5906,5907,5908,5938,5939,5970,5971,6002,6003,6068,6069,6071,6072,6073,6074,6075,6076,6077,6086,6089,6090,6091,6092,6093,6094,6095,6096,6097,6098,6099,6109,6155,6156,6157,6159,6277,6278,6313,6432,6433,6434,6439,6440,6450,6457,6458,6459,6679,6680,6683,6742,6744,6745,6746,6747,6748,6749,6750,6752,6754,6757,6758,6759,6760,6761,6762,6763,6764,6771,6772,6773,6774,6775,6776,6777,6778,6779,6780,6783,6832,6833,6834,6835,6836,6837,6838,6839,6840,6841,6842,6843,6844,6845,6846,6847,6848,6849,6850,6851,6852,6853,6854,6855,6856,6857,6858,6859,6860,6861,6862,6912,6913,6914,6915,6964,6966,6967,6968,6969,6970,6972,6978,7019,7020,7021,7022,7023,7024,7025,7026,7027,7040,7041,7074,7075,7076,7077,7080,7081,7083,7084,7085,7142,7144,7145,7149,7151,7152,7153,7212,7213,7214,7215,7216,7217,7218,7219,7222,7223,7376,7377,7378,7380,7381,7382,7383,7384,7385,7386,7387,7388,7389,7390,7391,7392,7394,7395,7396,7397,7398,7399,7400,7405,7412,7416,7417,7616,7617,7618,7619,7620,7621,7622,7623,7624,7625,7626,7627,7628,7629,7630,7631,7632,7633,7634,7635,7636,7637,7638,7639,7640,7641,7642,7643,7644,7645,7646,7647,7648,7649,7650,7651,7652,7653,7654,7655,7656,7657,7658,7659,7660,7661,7662,7663,7664,7665,7666,7667,7668,7669,7670,7671,7672,7673,7674,7675,7676,7677,7678,7679,8203,8206,8207,8234,8235,8236,8237,8238,8288,8289,8290,8291,8292,8298,8299,8300,8301,8302,8303,8400,8401,8402,8403,8404,8405,8406,8407,8408,8409,8410,8411,8412,8413,8414,8415,8416,8417,8418,8419,8420,8421,8422,8423,8424,8425,8426,8427,8428,8429,8430,8431,8432,11503,11504,11505,11647,11744,11745,11746,11747,11748,11749,11750,11751,11752,11753,11754,11755,11756,11757,11758,11759,11760,11761,11762,11763,11764,11765,11766,11767,11768,11769,11770,11771,11772,11773,11774,11775,12330,12331,12332,12333,12441,12442,42607,42608,42609,42610,42612,42613,42614,42615,42616,42617,42618,42619,42620,42621,42654,42655,42736,42737,43010,43014,43019,43045,43046,43052,43204,43205,43232,43233,43234,43235,43236,43237,43238,43239,43240,43241,43242,43243,43244,43245,43246,43247,43248,43249,43263,43302,43303,43304,43305,43306,43307,43308,43309,43335,43336,43337,43338,43339,43340,43341,43342,43343,43344,43345,43392,43393,43394,43443,43446,43447,43448,43449,43452,43453,43493,43561,43562,43563,43564,43565,43566,43569,43570,43573,43574,43587,43596,43644,43696,43698,43699,43700,43703,43704,43710,43711,43713,43756,43757,43766,44005,44008,44013,64286,65024,65025,65026,65027,65028,65029,65030,65031,65032,65033,65034,65035,65036,65037,65038,65039,65056,65057,65058,65059,65060,65061,65062,65063,65064,65065,65066,65067,65068,65069,65070,65071,65279,65529,65530,65531,66045,66272,66422,66423,66424,66425,66426,68097,68098,68099,68101,68102,68108,68109,68110,68111,68152,68153,68154,68159,68325,68326,68900,68901,68902,68903,69291,69292,69373,69374,69375,69446,69447,69448,69449,69450,69451,69452,69453,69454,69455,69456,69506,69507,69508,69509,69633,69688,69689,69690,69691,69692,69693,69694,69695,69696,69697,69698,69699,69700,69701,69702,69744,69747,69748,69759,69760,69761,69811,69812,69813,69814,69817,69818,69826,69888,69889,69890,69927,69928,69929,69930,69931,69933,69934,69935,69936,69937,69938,69939,69940,70003,70016,70017,70070,70071,70072,70073,70074,70075,70076,70077,70078,70089,70090,70091,70092,70095,70191,70192,70193,70196,70198,70199,70206,70209,70367,70371,70372,70373,70374,70375,70376,70377,70378,70400,70401,70459,70460,70464,70502,70503,70504,70505,70506,70507,70508,70512,70513,70514,70515,70516,70712,70713,70714,70715,70716,70717,70718,70719,70722,70723,70724,70726,70750,70835,70836,70837,70838,70839,70840,70842,70847,70848,70850,70851,71090,71091,71092,71093,71100,71101,71103,71104,71132,71133,71219,71220,71221,71222,71223,71224,71225,71226,71229,71231,71232,71339,71341,71344,71345,71346,71347,71348,71349,71351,71453,71454,71455,71458,71459,71460,71461,71463,71464,71465,71466,71467,71727,71728,71729,71730,71731,71732,71733,71734,71735,71737,71738,71995,71996,71998,72003,72148,72149,72150,72151,72154,72155,72160,72193,72194,72195,72196,72197,72198,72199,72200,72201,72202,72243,72244,72245,72246,72247,72248,72251,72252,72253,72254,72263,72273,72274,72275,72276,72277,72278,72281,72282,72283,72330,72331,72332,72333,72334,72335,72336,72337,72338,72339,72340,72341,72342,72344,72345,72752,72753,72754,72755,72756,72757,72758,72760,72761,72762,72763,72764,72765,72767,72850,72851,72852,72853,72854,72855,72856,72857,72858,72859,72860,72861,72862,72863,72864,72865,72866,72867,72868,72869,72870,72871,72874,72875,72876,72877,72878,72879,72880,72882,72883,72885,72886,73009,73010,73011,73012,73013,73014,73018,73020,73021,73023,73024,73025,73026,73027,73028,73029,73031,73104,73105,73109,73111,73459,73460,73472,73473,73526,73527,73528,73529,73530,73536,73538,78896,78897,78898,78899,78900,78901,78902,78903,78904,78905,78906,78907,78908,78909,78910,78911,78912,78919,78920,78921,78922,78923,78924,78925,78926,78927,78928,78929,78930,78931,78932,78933,92912,92913,92914,92915,92916,92976,92977,92978,92979,92980,92981,92982,94031,94095,94096,94097,94098,94180,113821,113822,113824,113825,113826,113827,118528,118529,118530,118531,118532,118533,118534,118535,118536,118537,118538,118539,118540,118541,118542,118543,118544,118545,118546,118547,118548,118549,118550,118551,118552,118553,118554,118555,118556,118557,118558,118559,118560,118561,118562,118563,118564,118565,118566,118567,118568,118569,118570,118571,118572,118573,118576,118577,118578,118579,118580,118581,118582,118583,118584,118585,118586,118587,118588,118589,118590,118591,118592,118593,118594,118595,118596,118597,118598,119143,119144,119145,119155,119156,119157,119158,119159,119160,119161,119162,119163,119164,119165,119166,119167,119168,119169,119170,119173,119174,119175,119176,119177,119178,119179,119210,119211,119212,119213,119362,119363,119364,121344,121345,121346,121347,121348,121349,121350,121351,121352,121353,121354,121355,121356,121357,121358,121359,121360,121361,121362,121363,121364,121365,121366,121367,121368,121369,121370,121371,121372,121373,121374,121375,121376,121377,121378,121379,121380,121381,121382,121383,121384,121385,121386,121387,121388,121389,121390,121391,121392,121393,121394,121395,121396,121397,121398,121403,121404,121405,121406,121407,121408,121409,121410,121411,121412,121413,121414,121415,121416,121417,121418,121419,121420,121421,121422,121423,121424,121425,121426,121427,121428,121429,121430,121431,121432,121433,121434,121435,121436,121437,121438,121439,121440,121441,121442,121443,121444,121445,121446,121447,121448,121449,121450,121451,121452,121461,121476,121499,121500,121501,121502,121503,121505,121506,121507,121508,121509,121510,121511,121512,121513,121514,121515,121516,121517,121518,121519,122880,122881,122882,122883,122884,122885,122886,122888,122889,122890,122891,122892,122893,122894,122895,122896,122897,122898,122899,122900,122901,122902,122903,122904,122907,122908,122909,122910,122911,122912,122913,122915,122916,122918,122919,122920,122921,122922,123023,123184,123185,123186,123187,123188,123189,123190,123566,123628,123629,123630,123631,124140,124141,124142,124143,125136,125137,125138,125139,125140,125141,125142,125252,125253,125254,125255,125256,125257,125258,125259,917505,917536,917537,917538,917539,917540,917541,917542,917543,917544,917545,917546,917547,917548,917549,917550,917551,917552,917553,917554,917555,917556,917557,917558,917559,917560,917561,917562,917563,917564,917565,917566,917567,917568,917569,917570,917571,917572,917573,917574,917575,917576,917577,917578,917579,917580,917581,917582,917583,917584,917585,917586,917587,917588,917589,917590,917591,917592,917593,917594,917595,917596,917597,917598,917599,917600,917601,917602,917603,917604,917605,917606,917607,917608,917609,917610,917611,917612,917613,917614,917615,917616,917617,917618,917619,917620,917621,917622,917623,917624,917625,917626,917627,917628,917629,917630,917631,917760,917761,917762,917763,917764,917765,917766,917767,917768,917769,917770,917771,917772,917773,917774,917775,917776,917777,917778,917779,917780,917781,917782,917783,917784,917785,917786,917787,917788,917789,917790,917791,917792,917793,917794,917795,917796,917797,917798,917799,917800,917801,917802,917803,917804,917805,917806,917807,917808,917809,917810,917811,917812,917813,917814,917815,917816,917817,917818,917819,917820,917821,917822,917823,917824,917825,917826,917827,917828,917829,917830,917831,917832,917833,917834,917835,917836,917837,917838,917839,917840,917841,917842,917843,917844,917845,917846,917847,917848,917849,917850,917851,917852,917853,917854,917855,917856,917857,917858,917859,917860,917861,917862,917863,917864,917865,917866,917867,917868,917869,917870,917871,917872,917873,917874,917875,917876,917877,917878,917879,917880,917881,917882,917883,917884,917885,917886,917887,917888,917889,917890,917891,917892,917893,917894,917895,917896,917897,917898,917899,917900,917901,917902,917903,917904,917905,917906,917907,917908,917909,917910,917911,917912,917913,917914,917915,917916,917917,917918,917919,917920,917921,917922,917923,917924,917925,917926,917927,917928,917929,917930,917931,917932,917933,917934,917935,917936,917937,917938,917939,917940,917941,917942,917943,917944,917945,917946,917947,917948,917949,917950,917951,917952,917953,917954,917955,917956,917957,917958,917959,917960,917961,917962,917963,917964,917965,917966,917967,917968,917969,917970,917971,917972,917973,917974,917975,917976,917977,917978,917979,917980,917981,917982,917983,917984,917985,917986,917987,917988,917989,917990,917991,917992,917993,917994,917995,917996,917997,917998,917999]
-};
